refactored validateOnSubmit method
